feat(礼包管理): 兑换记录

+++ backend/app/admin/models/item.go
@@ -8,7 +8,7 @@ import (
 type Item struct {
 	models.Model
 
-	ItemId   int64         `json:"itemId" gorm:"type:bigint;comment:道具id"`
+	ItemId   int           `json:"itemId" gorm:"type:bigint;comment:道具id"`
 	Type     int64         `json:"type" gorm:"type:bigint;comment:道具类型"`
 	Name     string        `json:"name" gorm:"type:varchar(8);comment:道具名称"`
 	Desc     string        `json:"desc" gorm:"type:varchar(255);comment:道具描述"`
